#6: Maintain local row array
Instead maintaining a dict with card ID (we do not know before the
update for the one on the Trello board only), we maintain a array with
all the spreadsheet cards, then after the update, all the cards in the
lists that are not in this array (via the ID comparison) are appended.

The ID is updated/created as we are using the same instances.